# Optimization: -1 is extra computation of string length internally