./parse.zsh "zplugin.zsh" > out_norm.txt  106,17s user 7,84s system 98% cpu 1:55,85 total

./parse.zsh "zplugin.zsh" > out_opt.txt  51,91s user 7,44s system 99% cpu 59,912 total

% wc -l zplugin.zsh
3188 zplugin.zsh

Tested with extracted main parser:
https://github.com/psprint/zsh-tools/blob/master/parse.zsh